pyproject.toml
[project]
name = "prototype-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3[crt]>=1.42.30",
    "fastapi>=0.128.0",
    "mangum>=0.20.0",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.13",
]
